[Enhancement] Schemachange can convert from latest version for PK table (#31376)
Fixes #31128

Currently, schema change may fail if it executes too long and the specified version may GCed(default 30 minutes), this means schema change that lasts over 30 minutes are likely to fail, which often happens for huge tables, wastes a lot of resources, and fails to get the job done.

This PR changes schema change behavior to convert from the base tablet's latest version rather than the specified version, avoiding the version GCed problem.

This PR also split UTs in tablet_updates_test, cause it's too large.
